fix(desktop): split AppShell notification effects
Move desktop notification callbacks and action handling into a focused AppShell hook so AppShell stays below the file-size guard without relying on whitespace churn.
